+++ RUFAS/simulation_engine.py
@@ -70,8 +70,8 @@ class SimulationEngine:
             self.herd_manager.herd_statistics,
             self.herd_manager.herd_reproduction_statistics,
             self.time,
-            self.herd_manager.heiferIIs,
-            self.herd_manager.cows,
+            self.herd_manager.heiferII_events_by_id,
+            self.herd_manager.cow_events_by_id,
         )
         # EEEManager.estimate_all()
         t_end_sim = timer.time()
@@ -167,8 +167,7 @@ class SimulationEngine:
         )
         self.feed_manager.manage_ration_interval_purchases(requested_feed, self.time)
 
-        for pen in self.herd_manager.all_pens:
-            AnimalModuleReporter.report_ration_interval_data(pen, self.time.simulation_day)
+        self.herd_manager.report_ration_interval_data(self.time.simulation_day)
 
         self.feed_manager.report_feed_manager_balance(self.time.simulation_day)